+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## 07/12/21
 After accidentally deleting all my files, I decided to restart from the start to make sure i do this right.
 * The Initial Commit of most of my files
+* Attempted to get pip install -e. to work, but failed. Will try again tomorrow
 
-## 08/12/21
-*
+## 08/12/21, 16:19
+* Added Logging System